Added a layer for buildings inside gui windows. Texture atlas now works with layers instead of indices.

+++ Freeserf.Core/Map.cs
@@ -673,7 +673,7 @@ namespace Freeserf
                     (uint)(virtualScreenWidth / tileWidth),
                     (uint)(virtualScreenHeight / tileHeight),
                     this, renderView.TriangleFactory,
-                    Render.TextureAtlasManager.Instance.GetOrCreate((int)Layer.Landscape),
+                    Render.TextureAtlasManager.Instance.GetOrCreate(Layer.Landscape),
                     dataSource);
             }
